Separate sign up page into own component
Ref #2
